Add path to Database and make elapsed an int.

+++ src/Timer.h
@@ -12,29 +12,29 @@ public:
     {
     }
 
-    long int start()
+    int start()
     {
         return (mStart = current());
     }
 
-    long int startTime() const
+    int startTime() const
     {
         return mStart;
     }
 
-    static long int current()
+    static int current()
     {
         timeval t;
         ::gettimeofday(&t, 0);
         return (t.tv_sec * 1000) + (t.tv_usec / 1000);
     }
 
-    long int elapsed() const
+    int elapsed() const
     {
         return current() - mStart;
     }
 
-    long int restart()
+    int restart()
     {
         const long int cur = current();
         const long int ret = cur - mStart;
